Module Bos_setup.Char

include module type of struct include Astring.Char end

Bytes

type t = char

The type for bytes.

val of_byte : int -> char

of_byte b is a byte from b.

raises Invalid_argument
if b is not in the range [0x00;0xFF].

val of_int : int -> char option

of_int b is a byte from b. None is returned if b is not in the range [0x00;0xFF].

val to_int : char -> int

to_int b is the byte b as an integer.

val hash : char -> int

hash is Hashtbl.hash.

Predicates

val equal : char -> char -> bool

equal b b' is b = b'.

val compare : char -> char -> int

compare b b' is Stdlib.compare b b'.

Bytes as US-ASCII characters

module Ascii = Astring.Char.Ascii

US-ASCII character support

Pretty printing

val pp : Format.formatter -> char -> unit

pp ppf c prints c on ppf.

val dump : Format.formatter -> char -> unit

dump ppf c prints c as a syntactically valid OCaml char on ppf using Ascii.escape_char