Module Bos_setup.OS

include module type of struct include Bos.OS end

Results

The functions of this module never raise Sys_error or Unix.Unix_error instead they turn these errors into error messages. If you need fine grained control over unix errors use the lower level functions in Bos.OS.U.

type ('a, 'e) result = ('a, [> Rresult.R.msg ] as 'e) result

The type for OS results.

Environment variables and program arguments

module Env = Bos.OS.Env

Environment variables.

module Arg = Bos.OS.Arg

Quick and dirty program arguments parsing.

File system operations

Note. When paths are relative they are expressed relative to the current working directory.

module Path = Bos.OS.Path

Path operations.

module File = Bos.OS.File

File operations.

module Dir = Bos.OS.Dir

Directory operations.

Commands

module Cmd = Bos.OS.Cmd

Executing commands.

Low level Unix access

module U = Bos.OS.U

Low level Unix access.